requirements.txt
Flask~=1.0.2
grpcio~=1.14.2
Keras~=2.2.2
Keras-Applications~=1.0.5
Keras-Preprocessing~=1.0.3
matplotlib~=2.2.3
numpy~=1.15.1
Pillow~=5.2.0
scikit-image~=0.14.1
scipy~=1.1.0
theano~=1.0.3
termcolor~=1.1.0